requirements.txt
django
pillow
djangorestframework
djangorestframework-simplejwt
django-filter
markdown
pymysql
python-dotenv
django-rosetta
django-cors-headers
PyJWT
